@@ -2,7 +2,7 @@ namespace Shared.Models;
 
 public class PagedList<T>
 {
-    public List<T> Data { get; set; } = new List<T>();
+    public List<T> Data { get; set; } = [];
     public int TotalCount { get; set; }
     public int PageNumber { get; set; }
     public int PageSize { get; set; }
