@@ -1,37 +1,44 @@
 using Microsoft.EntityFrameworkCore;
 using Server.Data;
 
-var builder = WebApplication.CreateBuilder(args);
+namespace Server;
 
-builder.Services.AddControllers();
-
-builder.Services.AddDbContext<PokemonDbContext>(options =>
-    // From appsettings.json
-    options.UseSqlServer(builder.Configuration.GetConnectionString("DefaultConnection"))); 
-
-builder.Services.AddLogging();
-builder.Services.AddOpenApi();
-
-// Add CORS to allow your Blazor client to make requests
-builder.Services.AddCors(options =>
+public abstract class Program
 {
-    options.AddPolicy("BlazorClientPolicy", policy =>
+    public static void Main (string[] args)
     {
-        policy.WithOrigins("https://localhost:7292") // Client port number
-            .AllowAnyMethod()
-            .AllowAnyHeader();
-    });
-});
+        var builder = WebApplication.CreateBuilder(args);
 
-var app = builder.Build();
+        builder.Services.AddControllers();
 
-if (app.Environment.IsDevelopment())
-{
-    app.MapOpenApi();
-    app.UseDeveloperExceptionPage();
+        builder.Services.AddDbContext<PokemonDbContext>(options =>
+            options.UseSqlite("Data Source=Data/PokemonDb.db"));
+
+        builder.Services.AddLogging();
+        builder.Services.AddOpenApi();
+
+        // Add CORS to allow your Blazor client to make requests
+        builder.Services.AddCors(options =>
+        {
+            options.AddPolicy("BlazorClientPolicy", policy =>
+            {
+                policy.WithOrigins("https://localhost:7292") // Client port number
+                    .AllowAnyMethod()
+                    .AllowAnyHeader();
+            });
+        });
+
+        var app = builder.Build();
+
+        if (app.Environment.IsDevelopment())
+        {
+            app.MapOpenApi();
+            app.UseDeveloperExceptionPage();
+        }
+
+        app.UseHttpsRedirection();
+        app.UseCors("BlazorClientPolicy");
+        app.MapControllers();
+        app.Run();
+    }
 }
-
-app.UseHttpsRedirection();
-app.UseCors("BlazorClientPolicy");
-app.MapControllers();
-app.Run();
